refactor: move `performAuthorizationCodeFlow` into the okta lib folder
