+++ src/server/routes/signIn.ts
@@ -15,18 +15,11 @@ import { ApiError } from '@/server/models/Error';
 import { typedRouter as router } from '@/server/lib/typedRoutes';
 import { readEmailCookie } from '@/server/lib/emailCookie';
 import handleRecaptcha from '@/server/lib/recaptcha';
-import {
-  generateAuthorizationState,
-  ProfileOpenIdClient,
-  setAuthorizationStateCookie,
-} from '@/server/lib/okta/openid-connect';
 import { OktaError } from '@/server/models/okta/Error';
 import {
   readEncryptedStateCookie,
   updateEncryptedStateCookie,
 } from '@/server/lib/encryptedStateCookie';
-import { getPersistableQueryParams } from '@/shared/lib/queryParams';
-import { RoutePaths } from '@/shared/model/Routes';
 import { CONSENTS_POST_SIGN_IN_PAGE } from '@/shared/model/Consent';
 import {
   getUserConsentsForPage,
@@ -37,50 +30,9 @@ import { loginMiddleware } from '@/server/lib/middleware/login';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { _ } from '@guardian/libs/dist/types/logger';
 import postSignInController from '@/server/lib/postSignInController';
+import { performAuthorizationCodeFlow } from '@/server/lib/okta/oauth';
 
 const { okta } = getConfiguration();
-
-/**
- * Method to perform the Authorization Code Flow
- * for a) the sign in session check
- * and b) post authentication (with the session token)
- * @param res - the express response object
- * @param sessionToken (optional) - if provided, we'll use this to set the session cookie
- * @param confirmationPagePath (optional) - page to redirect the user to after authentication
- * @returns 303 redirect to the okta /authorize endpoint
- */
-export const performAuthCodeFlow = (
-  res: ResponseWithRequestState,
-  sessionToken?: string,
-  confirmationPagePath?: RoutePaths,
-) => {
-  // firstly we generate and store a "state"
-  // as a http only, secure, signed session cookie
-  // which is a json object that contains a stateParam and the query params
-  // the stateParam is used to protect against csrf
-  const authState = generateAuthorizationState(
-    getPersistableQueryParams(res.locals.queryParams),
-    confirmationPagePath,
-  );
-  setAuthorizationStateCookie(authState, res);
-
-  // generate the /authorize endpoint url which we'll redirect the user too
-  const authorizeUrl = ProfileOpenIdClient.authorizationUrl({
-    // Don't prompt for authentication or consent
-    prompt: 'none',
-    // The sessionToken from authentication to exchange for session cookie
-    sessionToken,
-    // we send the generated stateParam as the state parameter
-    state: authState.stateParam,
-    // any scopes, by default the 'openid' scope is required
-    // the idapi_token_cookie_exchange scope is checked on IDAPI to return
-    // idapi cookies on authentication
-    scope: 'openid idapi_token_cookie_exchange',
-  });
-
-  // redirect the user to the /authorize endpoint
-  return res.redirect(303, authorizeUrl);
-};
 
 /**
  * Helper method to determine if a global error should show on the sign in page
@@ -158,7 +110,7 @@ router.get(
       } else {
         // sign in session check is not complete, so we want to check if a session exists
         // we do this by making an auth code request to okta
-        return performAuthCodeFlow(res);
+        return performAuthorizationCodeFlow(res);
       }
     } else {
       return showSignInPage(req, res);
@@ -259,7 +211,7 @@ const oktaSignInController = async (
     // we now need to generate an okta session
     // so we'll call the OIDC /authorize endpoint which sets a session cookie
     // we'll pretty much be performing the Authorization Code Flow
-    return performAuthCodeFlow(res, response.sessionToken);
+    return performAuthorizationCodeFlow(res, response.sessionToken);
   } catch (error) {
     trackMetric('OktaSignIn::Failure');
